# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "name": "Harsh Mittal",
            "email": "harshmittal2210@gmail.com"
        }
    },
    "paths": {
...
        "/accounts/": {
            "get": {
                "description": "get all bank account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bank Accounts"
                ],
                "summary": "Get All Account",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page Number",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Items Per Page",
                        "name": "items_per_page",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_bytesByHarsh_go-my-info_models.PaginatedListResp-models_BankAccount"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.JSONerrResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "create new bank account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bank Accounts"
                ],
                "summary": "Create Account",
                "parameters": [
                    {
                        "description": "Bank Account Body",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddBankAccountReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.BankAccount"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.JSONerrResponse"
                        }
                    }
                }
            }
        },
        "/accounts/{account_id}": {
            "get": {
                "description": "get bank account details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bank Accounts"
                ],
                "summary": "Get Account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bank Account ID",
                        "name": "account_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.BankAccount"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.JSONerrResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "update bank account details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bank Accounts"
                ],
                "summary": "Update Account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bank Account ID",
                        "name": "account_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Bank Account Body",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateBankAccountReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.JSONerrResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.JSONerrResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_bytesByHarsh_go-my-info_models.PaginatedListResp-models_BankAccount": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.BankAccount"
                    }
                },
                "has_more": {
                    "type": "boolean"
                },
                "items_per_page": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
...
        "models.AddBankAccountReq": {
            "type": "object",
            "required": [
                "account_number",
                "account_type",
                "balance",
                "bank_id",
                "currency",
                "name"
            ],
            "properties": {
                "account_number": {
                    "type": "string"
                },
                "account_type": {
                    "type": "string"
                },
                "balance": {
                    "type": "string"
                },
                "bank_id": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.BankAccount": {
            "type": "object",
            "properties": {
                "account_number": {
                    "type": "string"
                },
                "account_type": {
                    "type": "string"
                },
                "balance": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.JSONerrResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "models.UpdateBankAccountReq": {
            "type": "object",
            "required": [
                "account_number",
                "account_type",
                "balance",
                "currency",
                "name"
            ],
            "properties": {
                "account_number": {
                    "type": "string"
                },
                "account_type": {
                    "type": "string"
                },
                "balance": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    }
}